newline problem Hey Onais,
Can you please have a look at this test case?

go run . --output=banner.txt "hello\nfirst" standard

apparently the newline is not being handled properly.  I have been trying to fix it but not succeded
the PatternsToArr func is responsable for  taking a string returning an array by splititng it a newline.